Make thread pool header-only.
